# Faster implementation of fANOVA