--- a/optuna/importance/_fanova/_tree.py
+++ b/optuna/importance/_fanova/_tree.py
@@ -1,3 +1,4 @@
+from functools import lru_cache
 import itertools
 from typing import List
 from typing import Set
@@ -93,7 +94,7 @@ class _FanovaTree:
         active_search_spaces = [search_spaces]
 
         node_indices = []
-        active_features_cardinalities = []
+        active_leaf_search_spaces = []
 
         while len(active_nodes) > 0:
             node_index = active_nodes.pop()
@@ -120,7 +121,7 @@ class _FanovaTree:
                     continue
 
                 # If subtree starting from node splits on an active feature, push both child nodes.
-                if (active_features & self._subtree_active_features[node_index]).any():
+                if any(self._subtree_active_features[node_index][active_features].tolist()):
                     for child_node_index in self._get_node_children(node_index):
                         active_nodes.append(child_node_index)
                         active_search_spaces.append(search_spaces)
@@ -128,11 +129,16 @@ class _FanovaTree:
 
             # If node is a leaf or the subtree does not split on any of the active features.
             node_indices.append(node_index)
-            active_features_cardinalities.append(_get_cardinality(search_spaces))
+            active_leaf_search_spaces.append(search_spaces)
 
         statistics = self._statistics[node_indices]
         values = statistics[:, 0]
         weights = statistics[:, 1]
+        active_features_cardinalities = numpy.asarray(active_leaf_search_spaces)
+        active_features_cardinalities = (
+            active_features_cardinalities[:, :, 1] - active_features_cardinalities[:, :, 0]
+        )
+        active_features_cardinalities = numpy.prod(active_features_cardinalities, axis=1)
         weights = weights / active_features_cardinalities
 
         value = numpy.average(values, weights=weights)
@@ -242,24 +248,31 @@ class _FanovaTree:
     def _n_nodes(self) -> int:
         return self._tree.node_count
 
+    @lru_cache(maxsize=None)
     def _is_node_leaf(self, node_index: int) -> bool:
         return self._tree.feature[node_index] < 0
 
+    @lru_cache(maxsize=None)
     def _get_node_left_child(self, node_index: int) -> int:
         return self._tree.children_left[node_index]
 
+    @lru_cache(maxsize=None)
     def _get_node_right_child(self, node_index: int) -> int:
         return self._tree.children_right[node_index]
 
+    @lru_cache(maxsize=None)
     def _get_node_children(self, node_index: int) -> Tuple[int, int]:
         return self._get_node_left_child(node_index), self._get_node_right_child(node_index)
 
+    @lru_cache(maxsize=None)
     def _get_node_value(self, node_index: int) -> float:
         return float(self._tree.value[node_index])
 
+    @lru_cache(maxsize=None)
     def _get_node_split_threshold(self, node_index: int) -> float:
         return self._tree.threshold[node_index]
 
+    @lru_cache(maxsize=None)
     def _get_node_split_feature(self, node_index: int) -> int:
         return self._tree.feature[node_index]