## Motivation
This PR refactors the implementation of `FanovaImportanceEvaluator` to improve speed.

## Description of the changes
- Adding `lru_cache` to node information getters
- Discarding NumPy's `any`
- Reducing the number of cardinality calculations

## Benchmark
I ran a benchmark using the optuna-fast-fanova's script. https://github.com/optuna/optuna-fast-fanova/blob/main/tools/benchmark.py

| n_trials | n_params | n_trees | Optuna (master) | Optuna (PR) | optuna-fast-fanova |
| -------- | -------- | ------- | --------------- | ------------------ | ------------------ |
| 100 | 2 | 64 | 1.219s | 0.844s (-30.8%) | 0.111s (-90.9%) |
| 100 | 8 | 64 | 1.611s | 1.031s (-36.0%) | 0.170s (-89.5%) |
| 100 | 32 | 64 | 1.763s | 1.245s (-29.4%) | 0.438s (-75.2%) |
| 1000 | 2 | 64 | 51.800s | 24.699s (-52.3%) | 1.359s (-97.4%) |
| 1000 | 8 | 64 | 79.404s | 34.244s (-56.9%) | 2.342s (-97.0%) |
| 1000 | 32 | 64 | 58.464s | 26.081s (-55.4%) | 3.464s (-94.1%) |